Implement old data clear function
