+++ algo/firebase/front/src/components/000_menu/00_menu/00_menuMain.tsx
@@ -1,52 +1,27 @@
-import { useRecoilState, useSetRecoilState } from 'recoil';
+import { useRecoilState } from 'recoil';
 
 import Header from '@/components/common/Header';
-import {
-  residentInputState,
-  residentVCListState,
-  residentVCRequestListState,
-  accountInputState,
-  accountVCListState,
-  accountVCRequestListState,
-  taxInputState,
-  taxVCListState,
-  taxVCRequestListState,
-  subsidyInputState,
-  subsidyListState,
-  VCListState
-} from '@/lib/states/mockApp';
 
 import holderDidAccountState from '@/lib/states/holderDidAccountState';
 import issuerDidAccountState from '@/lib/states/issuerDidAccountState';
 import verifierDidAccountState from '@/lib/states/verifierDidAccountState';
 import { useEffect, useState } from 'react';
 import SelectActorButton from '../../common/SelectActorButton';
-import { useErrorHandler } from 'react-error-boundary';
+import useDataClear from '@/components/util/useDataClear';
 
 
 
 const MenuMain = () => {
-  const clearResidentInputState = useSetRecoilState(residentInputState);
-  const clearResidentVCListState = useSetRecoilState(residentVCListState);
-  const clearResidentVCRequestListState = useSetRecoilState(residentVCRequestListState);
-  const clearAccountInputState = useSetRecoilState(accountInputState);
-  const clearAccountVCListState = useSetRecoilState(accountVCListState);
-  const clearAccountVCRequestListState = useSetRecoilState(accountVCRequestListState);
-  const clearTaxInputState = useSetRecoilState(taxInputState);
-  const clearTaxVCListState = useSetRecoilState(taxVCListState);
-  const clearTaxVCRequestListState = useSetRecoilState(taxVCRequestListState);
-  const clearSubsidyInputState = useSetRecoilState(subsidyInputState);
-  const clearSubsidyListState = useSetRecoilState(subsidyListState);
-  const clearVCListState = useSetRecoilState(VCListState);
+
 
   const [holderDidAccountGlobal] = useRecoilState(holderDidAccountState);
   const [issuerDidAccountGlobal] = useRecoilState(issuerDidAccountState);
   const [verifierDidAccountGlobal] = useRecoilState(verifierDidAccountState);
 
   const [haveDid, setHaveDid] = useState(true)
-  const [clearMsg, setclearMsg] = useState("")
 
-  const errorHandler = useErrorHandler();
+  const { clearMsg, clearAllState } = useDataClear();
+
 
   useEffect(() => {
     if (holderDidAccountGlobal && issuerDidAccountGlobal && verifierDidAccountGlobal) {
@@ -57,98 +32,6 @@ const MenuMain = () => {
     }
   }, [holderDidAccountGlobal, issuerDidAccountGlobal, verifierDidAccountGlobal]);
 
-  const handleClearInputState = () => {
-    clearResidentInputState(() => ({
-      id: 0,
-      fullName: '',
-      fullNameFurigana: '',
-      address: '',
-      addressRegistDate: '',
-      addressRegistYear: '',
-      addressRegistMonth: '',
-      permanentAddress: '',
-      applicationDate: undefined,
-      issueDate: undefined,
-      verifyStatus: undefined,
-      approvalStatus: undefined
-    }))
-    clearAccountInputState(() => ({
-      id: 0,
-      bankCode: '',
-      branchNumber: '',
-      accountNumber: '',
-      corporateName: '',
-      applicantName: '',
-      applicantAddress: '',
-      applicationDate: undefined,
-      issueDate: undefined,
-      verifyStatus: undefined,
-      approvalStatus: undefined,
-    }))
-    clearTaxInputState(() => ({
-      id: 0,
-      applicationYear: "",
-      corporationName: "",
-      corporationAddress: "",
-      fullName: "",
-      address: "",
-      applicationDate: "",
-      issueDate: "",
-      verifyStatus: false,
-      approvalStatus: false,
-    }))
-    clearSubsidyInputState(() => ({
-      id: 0,
-      resident: "",
-      account: "",
-      tax: "",
-      fullName: "",
-      address: "",
-      applicationDate: "",
-      issueDate: "",
-      residentVerifyStatus: false,
-      accountVerifyStatus: false,
-      taxVerifyStatus: false,
-      verifyStatus: false,
-      approvalStatus: false,
-      residentVP: undefined,
-      accountVP: undefined,
-      taxVP: undefined
-    }))
-  };
-
-
-
-  const onClear = () => {
-    try {
-      handleClearInputState();
-      clearResidentVCListState(() => []);
-      clearResidentVCRequestListState(() => []);
-      clearAccountVCListState(() => []);
-      clearAccountVCRequestListState(() => []);
-      clearTaxVCListState(() => []);
-      clearTaxVCRequestListState(() => []);
-      clearSubsidyListState(() => []);
-      clearVCListState(() => ({
-        resident: [],
-        account: [],
-        tax: [],
-        subsidy: []
-      }));
-      showClearSuccessMsg();
-    } catch (e) {
-      errorHandler(e);
-    }
-  };
-
-  const showClearSuccessMsg = async () => {
-    setclearMsg("データクリアが完了しました。")
-    await delay(5000);
-    setclearMsg("")
-  }
-
-  const delay = (ms: number | undefined) =>
-    new Promise((resolve) => setTimeout(resolve, ms));
 
   return (
     <>
@@ -175,7 +58,7 @@ const MenuMain = () => {
           <SelectActorButton target="subsidy" haveDid={haveDid} />
         </section>
         <div className="pt-32 pl-4 pb-4 flex justify-between">
-          <button onClick={onClear} className={"input-form-button-small " + (haveDid ? "" : "bg-color-gray-search opacity-30")} disabled={!haveDid}>
+          <button onClick={clearAllState} className={"input-form-button-small " + (haveDid ? "" : "bg-color-gray-search opacity-30")} disabled={!haveDid}>
             データクリア
           </button>
         </div>
